[InstCombine] ARM/ARM64: Fix ordering of AES instruction operands when combining with XOR.
The ARM aese/aesd instructions destructively place the result into
the first operand, so it's better to keep the data the first operand
rathan than key. However, when combining with a previous XOR
instruction we just assumed that the LHS is data, and RHS is key,
but when we get it wrong, the XOR is not eliminated, but get
replaced by a MOV instruction.

We can do a bit more to deduce which XOR operand is data by looking
for the operand which is output of a previous aesmc/aesimc.